+++ lib/algo_impl/eals/eals.cc
@@ -29,9 +29,9 @@ bool EALS::parse_option(std::string opt_path) {
     return ok;
 }
 
-void EALS::initialize_model(double* P_ptr,
-                            double* Q_ptr,
-                            double* C_ptr,
+void EALS::initialize_model(float* P_ptr,
+                            float* Q_ptr,
+                            float* C_ptr,
                             const int32_t P_rows,
                             const int32_t Q_rows) {
     const int32_t D = opt_["d"].int_value();
@@ -53,8 +53,8 @@ void EALS::precompute_cache(const int32_t nnz,
     if (is_cached) {
         return;
     }
-    const double* P = axis == 0 ? P_ptr_ : Q_ptr_;
-    const double* Q = axis == 0 ? Q_ptr_ : P_ptr_;
+    const float* P = axis == 0 ? P_ptr_ : Q_ptr_;
+    const float* Q = axis == 0 ? Q_ptr_ : P_ptr_;
     auto& vhat_cache = axis == 0 ? vhat_cache_u_ : vhat_cache_i_;
     if (nnz != vhat_cache.size()) {
         vhat_cache.resize(nnz);
@@ -71,10 +71,10 @@ void EALS::precompute_cache(const int32_t nnz,
             TRACE("No data exists for {}", xidx);
             continue;
         }
-        const double* p_ptr = &P[xidx * D];
+        const float* p_ptr = &P[xidx * D];
         for (int64_t ind=beg; ind < end; ++ind) {
             const int32_t yidx = keys[ind];
-            const double* q_ptr = &Q[yidx * D];
+            const float* q_ptr = &Q[yidx * D];
             vhat_cache[ind] = std::inner_product(p_ptr, p_ptr + D, q_ptr, kZero);
             idx_xmajor[ind].set(yidx, xidx, ind);
         }
@@ -114,7 +114,7 @@ bool EALS::update(const int64_t* indptr,
     return is_cached;
 }
 
-std::pair<double, double> EALS::estimate_loss(const int32_t nnz,
+std::pair<float, float> EALS::estimate_loss(const int32_t nnz,
                                               const int64_t* indptr,
                                               const int32_t* keys,
                                               const float* vals,
@@ -126,12 +126,12 @@ std::pair<double, double> EALS::estimate_loss(const int32_t nnz,
     const bool is_cached = is_P_cached_ && is_Q_cached_;
     if (!is_cached) {
         TRACE("Compute cache first(P: {}, Q: {}). Empty data is returned.", is_P_cached_, is_Q_cached_);
-        return std::pair<double, double>(kZero, kZero);
+        return std::pair<float, float>(kZero, kZero);
     }
     const int32_t num_workers = opt_["num_workers"].int_value();
-    const double alpha = opt_["alpha"].number_value();
-    std::vector<double> feedbacks4tid(num_workers, kZero);
-    std::vector<double> mse4tid(num_workers, kZero);
+    const float alpha = opt_["alpha"].number_value();
+    std::vector<float> feedbacks4tid(num_workers, kZero);
+    std::vector<float> mse4tid(num_workers, kZero);
     auto& vhat_cache = axis == 0 ? vhat_cache_u_ : vhat_cache_i_;
     const int32_t end_loop = axis == 0 ? P_rows_ : Q_rows_;
     #pragma omp parallel for schedule(dynamic, 8)
@@ -141,30 +141,30 @@ std::pair<double, double> EALS::estimate_loss(const int32_t nnz,
         const int64_t end = indptr[xidx];
         for (int64_t ind=beg; ind < end; ++ind) {
             const int32_t yidx = keys[ind];
-            const double v = static_cast<double>(vals[ind]);
-            const double vhat = vhat_cache[ind];
-            const double error = (v - vhat);
+            const float v = vals[ind];
+            const float vhat = vhat_cache[ind];
+            const float error = (v - vhat);
             feedbacks4tid[tid] += (kOne + alpha * v) * error * error;
             const int32_t iidx = axis == 0 ? yidx : xidx;
             feedbacks4tid[tid] -= C_ptr_[iidx] * vhat * vhat; // To avoid duplication in a term of negative feedbacks.
             mse4tid[tid] += error * error;
         }
     }
-    const double squared_error = std::accumulate(mse4tid.begin(), mse4tid.end(), kZero);
-    double feedbacks = std::accumulate(feedbacks4tid.begin(), feedbacks4tid.end(), kZero);
+    const float squared_error = std::accumulate(mse4tid.begin(), mse4tid.end(), kZero);
+    float feedbacks = std::accumulate(feedbacks4tid.begin(), feedbacks4tid.end(), kZero);
 
     // Add L2 regularization terms
     const int32_t D = opt_["d"].int_value();
-    const double reg_u = opt_["reg_u"].number_value();
-    const double reg_i = opt_["reg_i"].number_value();
-    auto op = [](const double & init, const double & v) -> double {
+    const float reg_u = opt_["reg_u"].number_value();
+    const float reg_i = opt_["reg_i"].number_value();
+    auto op = [](const float & init, const float & v) -> float {
         return init + v * v;
     };
-    const double reg = reg_u * std::accumulate(P_ptr_, P_ptr_ + (P_rows_ * D), kZero, op) +
+    const float reg = reg_u * std::accumulate(P_ptr_, P_ptr_ + (P_rows_ * D), kZero, op) +
                        reg_i * std::accumulate(Q_ptr_, Q_ptr_ + (Q_rows_ * D), kZero, op);
 
     // Add negative feedbacks: sum_{u} sum_{i \in R_u} C_{i} * vHat_{u,i}^2
-    std::vector<double> CQ(Q_rows_ * D, kZero);
+    std::vector<float> CQ(Q_rows_ * D, kZero);
     #pragma omp parallel for schedule(dynamic, 8)
     for (int32_t iidx=0; iidx < Q_rows_; ++iidx) {
         const int32_t ridx = iidx * D;
@@ -172,53 +172,53 @@ std::pair<double, double> EALS::estimate_loss(const int32_t nnz,
             CQ[ridx + d] = std::sqrt(C_ptr_[iidx]) * Q_ptr_[ridx + d];
         }
     }
-    std::vector<double> Sp(D * D, kZero);
-    std::vector<double> Sq(D * D, kZero);
+    std::vector<float> Sp(D * D, kZero);
+    std::vector<float> Sq(D * D, kZero);
     blas::syrk("u", "t", D, P_rows_, kOne, P_ptr_, kZero, Sp.data());
     blas::syrk("u", "t", D, Q_rows_, kOne, CQ.data(), kZero, Sq.data());
     feedbacks += std::inner_product(Sp.data(), Sp.data() + (D * D), Sq.data(), kZero);
-    const double rmse = std::sqrt(squared_error / nnz);
-    const double loss = feedbacks + reg;
-    return std::pair<double, double>(rmse, loss);
+    const float rmse = std::sqrt(squared_error / nnz);
+    const float loss = feedbacks + reg;
+    return std::pair<float, float>(rmse, loss);
 }
 
 void EALS::update_P_(const int64_t* indptr,
                      const int32_t* keys,
                      const float* vals) {
     const int32_t D = opt_["d"].int_value();
-    std::vector<double> CQ(Q_rows_ * D, kZero);
+    std::vector<float> CQ(Q_rows_ * D, kZero);
     for (int32_t iidx=0; iidx < Q_rows_; ++iidx) {
-        const double sqrt_C = std::sqrt(C_ptr_[iidx]);
+        const float sqrt_C = std::sqrt(C_ptr_[iidx]);
         const int32_t ridx = iidx * D;
-        const double* q_ptr = &Q_ptr_[ridx];
-        double* cq_ptr = &CQ[ridx];
+        const float* q_ptr = &Q_ptr_[ridx];
+        float* cq_ptr = &CQ[ridx];
         std::transform(q_ptr, q_ptr + D, cq_ptr,
-            [sqrt_C](const double elem) -> double {
+            [sqrt_C](const float elem) -> float {
                 return sqrt_C * elem;
             }
         );
     }
-    std::vector<double> Sq(D * D, kZero);
+    std::vector<float> Sq(D * D, kZero);
     blas::syrk("u", "t", D, Q_rows_, kOne, CQ.data(), kZero, Sq.data());
-    const double alpha = opt_["alpha"].number_value();
-    const double reg_u = opt_["reg_u"].number_value();
+    const float alpha = opt_["alpha"].number_value();
+    const float reg_u = opt_["reg_u"].number_value();
     #pragma omp parallel for schedule(dynamic, 8)
     for (int32_t uidx=0; uidx < P_rows_; ++uidx) {
-        double* p_ptr = &P_ptr_[uidx * D];
+        float* p_ptr = &P_ptr_[uidx * D];
         const int64_t beg = uidx == 0 ? 0 : indptr[uidx - 1];
         const int64_t end = indptr[uidx];
         for (int32_t d=0; d < D; ++d) {
-            double numerator = kZero;
-            double denominator = kZero;
+            float numerator = kZero;
+            float denominator = kZero;
             for (int64_t ind=beg; ind < end; ++ind) {
                 const int32_t iidx = keys[ind];
-                const double v = static_cast<double>(vals[ind]);
-                const double* q_ptr = &Q_ptr_[iidx * D];
-                const double vhat = vhat_cache_u_[ind];
-                const double pq = p_ptr[d] * q_ptr[d];
-                const double vf = vhat - pq;
-                const double w = (kOne + alpha * v);
-                const double wmc = w - C_ptr_[iidx];
+                const float v = vals[ind];
+                const float* q_ptr = &Q_ptr_[iidx * D];
+                const float vhat = vhat_cache_u_[ind];
+                const float pq = p_ptr[d] * q_ptr[d];
+                const float vf = vhat - pq;
+                const float w = (kOne + alpha * v);
+                const float wmc = w - C_ptr_[iidx];
                 numerator += (w * v - wmc * vf) * q_ptr[d];
                 denominator += wmc * q_ptr[d] * q_ptr[d];
                 vhat_cache_u_[ind] -= pq;
@@ -229,8 +229,8 @@ void EALS::update_P_(const int64_t* indptr,
             p_ptr[d] = numerator / denominator;
             for (int64_t ind=beg; ind < end; ++ind) {
                 const int32_t iidx = keys[ind];
-                const double* q_ptr = &Q_ptr_[iidx * D];
-                const double pq = p_ptr[d] * q_ptr[d];
+                const float* q_ptr = &Q_ptr_[iidx * D];
+                const float pq = p_ptr[d] * q_ptr[d];
                 vhat_cache_u_[ind] += pq;
                 vhat_cache_i_[ind_u2i_[ind]] += pq;
             }
@@ -242,27 +242,27 @@ void EALS::update_Q_(const int64_t* indptr,
                      const int32_t* keys,
                      const float* vals) {
     const int32_t D = opt_["d"].int_value();
-    std::vector<double> Sp(D * D, kZero);
+    std::vector<float> Sp(D * D, kZero);
     blas::syrk("u", "t", D, P_rows_, kOne, P_ptr_, kZero, Sp.data());
-    const double alpha = opt_["alpha"].number_value();
-    const double reg_i = opt_["reg_i"].number_value();
+    const float alpha = opt_["alpha"].number_value();
+    const float reg_i = opt_["reg_i"].number_value();
     #pragma omp parallel for schedule(dynamic, 8)
     for (int32_t iidx=0; iidx < Q_rows_; ++iidx) {
-        double* q_ptr = &Q_ptr_[iidx * D];
+        float* q_ptr = &Q_ptr_[iidx * D];
         const int64_t beg = iidx == 0 ? 0 : indptr[iidx - 1];
         const int64_t end = indptr[iidx];
         for (int32_t d=0; d < D; ++d) {
-            double numerator = kZero;
-            double denominator = kZero;
+            float numerator = kZero;
+            float denominator = kZero;
             for (int64_t ind=beg; ind < end; ++ind) {
                 const int32_t uidx = keys[ind];
-                const double v = static_cast<double>(vals[ind]);
-                const double* p_ptr = &P_ptr_[uidx * D];
-                const double vhat = vhat_cache_i_[ind];
-                const double pq = p_ptr[d] * q_ptr[d];
-                const double vf = vhat - pq;
-                const double w = (kOne + alpha * v);
-                const double wmc = w - C_ptr_[iidx];
+                const float v = vals[ind];
+                const float* p_ptr = &P_ptr_[uidx * D];
+                const float vhat = vhat_cache_i_[ind];
+                const float pq = p_ptr[d] * q_ptr[d];
+                const float vf = vhat - pq;
+                const float w = (kOne + alpha * v);
+                const float wmc = w - C_ptr_[iidx];
                 numerator += (w * v - wmc * vf) * p_ptr[d];
                 denominator += wmc * p_ptr[d] * p_ptr[d];
                 vhat_cache_i_[ind] -= pq;
@@ -273,8 +273,8 @@ void EALS::update_Q_(const int64_t* indptr,
             q_ptr[d] = numerator / denominator;
             for (int64_t ind=beg; ind < end; ++ind) {
                 const int32_t uidx = keys[ind];
-                const double* p_ptr = &P_ptr_[uidx * D];
-                const double pq = p_ptr[d] * q_ptr[d];
+                const float* p_ptr = &P_ptr_[uidx * D];
+                const float pq = p_ptr[d] * q_ptr[d];
                 vhat_cache_i_[ind] += pq;
                 vhat_cache_u_[ind_i2u_[ind]] += pq;
             }
